- Fixed a problem where the package was not imported correctly. - Fixed a problem where the tests were not comparing config data correctly and failing. - Update the readme. - Renamed the base package. - Updated some dependencies.
