+++ backup/backup.py
@@ -13,10 +13,11 @@ import subprocess
 import time
 import yaml
 from termcolor import cprint
-from src.data_classes import ConfigDataClass, ConfigListDataClass
-from src.exceptions import (PathValidationException, PathNotFoundException,
-                            RemoteSocketNotFoundException,
-                            ConfigJobNotFoundException, RsyncNotFoundException)
+from data_classes import ConfigDataClass, ConfigListDataClass
+from exceptions import (PathValidationException, PathNotFoundException,
+                        RemoteSocketNotFoundException,
+                        ConfigJobNotFoundException,
+                        RsyncNotFoundException)
 
 RSYNC_VERSION = '2.6.9'
 DEFAULT_CONFIG = 'config.yaml'
