+++ requirements.in
@@ -1,7 +1,8 @@
 flake8==5.0.4
-pip-tools==6.8.0
+pip-tools==6.9.0
 termcolor==2.0.1
-pytest==7.1.3
-pytest-cov==3.0.0
+pytest==7.2.0
+pytest-cov==4.0.0
 pytest-mock==3.10.0
 pyyaml==6.0
+build==0.9.0
