+++ requirements.txt
@@ -1,17 +1,21 @@
 #
-# This file is autogenerated by pip-compile with python 3.10
+# This file is autogenerated by pip-compile with python 3.9
 # To update, run:
 #
 #    pip-compile
 #
 attrs==22.1.0
     # via pytest
-build==0.8.0
-    # via pip-tools
+build==0.9.0
+    # via
+    #   -r requirements.in
+    #   pip-tools
 click==8.1.3
     # via pip-tools
-coverage[toml]==6.4.4
+coverage[toml]==6.5.0
     # via pytest-cov
+exceptiongroup==1.0.0
+    # via pytest
 flake8==5.0.4
     # via -r requirements.in
 iniconfig==1.1.1
@@ -24,11 +28,9 @@ packaging==21.3
     #   pytest
 pep517==0.13.0
     # via build
-pip-tools==6.8.0
+pip-tools==6.9.0
     # via -r requirements.in
 pluggy==1.0.0
-    # via pytest
-py==1.11.0
     # via pytest
 pycodestyle==2.9.1
     # via flake8
@@ -36,12 +38,12 @@ pyflakes==2.5.0
     # via flake8
 pyparsing==3.0.9
     # via packaging
-pytest==7.1.3
+pytest==7.2.0
     # via
     #   -r requirements.in
     #   pytest-cov
     #   pytest-mock
-pytest-cov==3.0.0
+pytest-cov==4.0.0
     # via -r requirements.in
 pytest-mock==3.10.0
     # via -r requirements.in
