+++ setup.cfg
@@ -2,8 +2,8 @@
 
 
 [metadata]
-name = RSYNC Backup Automation
-version = 2.0.1
+name = backup
+version = 2.1.0
 author = Mortolian
 maintainer = Mortolian
 keywords = rsync, backup, sync, ssh
@@ -14,7 +14,7 @@ classifiers =
 
 [options]
 package_dir =
-    =src
+    =backup
 packages = find:
 install_requires = file:requirements.txt
 zip_safe = False
@@ -22,7 +22,7 @@ python_requires = >=3.8
 
 
 [options.packages.find]
-where = src
+where = backup
 
 
 [flake8]
